refactor agent config
